PCI: Set the required config registers of function
- Set the required configuration registers according to PCIe technology.
- Use Vec structures instead of simple [0; u32].

+++ src/pci/src/constants.rs
@@ -2,22 +2,22 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /// There are up to 256 Bus numbers that can be assigned.
-pub const MAX_BUS_NUMBER: usize = 256;
+pub const MAX_BUS_NUMBER: u32 = 256;
 
 /// There are up to 32 Device attachments on a single PCI Bus.
-pub const MAX_DEVICE_NUMBER: usize = 32;
+pub const MAX_DEVICE_NUMBER: u32 = 32;
 
 /// A Device can have implemented up to 8 Functions (not necessarily sequentially).
-pub const MAX_FUNCTION_NUMBER: usize = 8;
+pub const MAX_FUNCTION_NUMBER: u32 = 8;
 
 /// The PCI Configuration Header Space has a length of 64 bytes, so 16 dwords.
-pub const CONFIGURATION_HEADER_SIZE: usize = 16;
+pub const CONFIGURATION_HEADER_SIZE: u32 = 16;
 
 /// The PCI optional registers (device specific) has a length of 192 bytes, so 48 dwords.
-pub const DEVICE_SPECIFIC_REGISTERS: usize = 48;
+pub const DEVICE_SPECIFIC_REGISTERS: u32 = 48;
 
 /// The PCIe Extended Configuration Registers Space has a length 3840 bytes, so 960 dwords.
-pub const EXTENDED_CONFIGURATION_REGISTERS: usize = 960;
+pub const EXTENDED_CONFIGURATION_REGISTERS: u32 = 960;
 
 /// Type 0 is required for every Function, except for the Bridge Functions.
 /// Type 1 is required for the Bridge (Switch) Functions.
